+++ db_operations/read_operations.go
@@ -40,7 +40,7 @@ func OpenFile(path string) *os.File {
 	return file
 }
 
-func Get(key string, index map[string]int64) (string, error) {
+func Get(key string, index map[string]int) (string, error) {
 	file := OpenFile("database/database.txt")
 	defer file.Close()
 
@@ -50,7 +50,7 @@ func Get(key string, index map[string]int64) (string, error) {
 		return "", nil
 	}
 
-	_, err := file.Seek(byteOffset, 0)
+	_, err := file.Seek(int64(byteOffset), 0)
 	if err != nil {
 		return "", err
 	}
